+++ lib/sisimai/reason/securityerror.rb
@@ -32,6 +32,7 @@ module Sisimai
           return nil unless argv1
           regex = %r{(?>
              Account[ ]not[ ]subscribed[ ]to[ ]SES
+            |Authentification[ ]requise.+[A-Z]{3}.+402
             |authentication[ ](?:
                Credentials Invalid
               |failure
@@ -40,6 +41,7 @@ module Sisimai
               |turned[ ]on[ ]in[ ]your[ ]email[ ]client
               )
             |\d+[ ]denied[ ]\[[a-z]+\][ ].+[(]Mode:[ ].+[)]
+            |Codes?[ ]d'?[ ]*authentification[ ]invalide.+[A-Z]{3}.+305
             |domain[ ].+[ ]is[ ]a[ ]dead[ ]domain
             |Executable[ ]files[ ]are[ ]not[ ]allowed[ ]in[ ]compressed[ ]files
             |insecure[ ]mail[ ]relay